requirements.txt
django-autoslug
djangorestframework
django
django-star-ratings
psycopg2
Pillow
